Update 04-functions.mkd
When we re-write the pay and grade computations as function calls in chapter 4, it seems like the text should show the function call instead of ask for the inputs, since we will put the inputs into the function call. Maybe I am interpreting this incorrectly. This book is a fantastic resource and I am learning a lot. Thanks!